Make: Test

+++ golang/lib.go
@@ -116,10 +116,10 @@ func NewIndex(conf IndexConfig) (index *Index, err error) {
 	index = &Index{config: conf}
 
 	conf = index.config
-	dimensions := C.ulong(conf.Dimensions)
-	connectivity := C.ulong(conf.Connectivity)
-	expansion_add := C.ulong(conf.ExpansionAdd)
-	expansion_search := C.ulong(conf.ExpansionSearch)
+	dimensions := C.size_t(conf.Dimensions)
+	connectivity := C.size_t(conf.Connectivity)
+	expansion_add := C.size_t(conf.ExpansionAdd)
+	expansion_search := C.size_t(conf.ExpansionSearch)
 	multi := C.bool(conf.Multi)
 
 	options := C.struct_usearch_init_options_t{}
